refactor: DepthwiseSeparableConvolution layer
